requirements.txt
# Core Dependencies
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
web3>=6.11.0

# Database
sqlalchemy>=2.0.23

# Notifications
python-telegram-bot>=20.7
discord.py>=2.3.2

# Twitter
tweepy>=4.14.0

# Data Analysis
numpy>=1.24.3
pandas>=2.0.3

# Utilities
schedule>=1.2.0
python-dateutil>=2.8.2

# Logging
colorlog>=6.8.0

# HTTP Client
httpx>=0.25.0
